fix(pipeline): move read model into factory.nats.pipeline package
Relocate pipeline modules under src/factory/nats/pipeline/ to satisfy
the 15-file folder size gate on factory/nats.
