+++ src/factory/nats/pipeline/__init__.py
@@ -1,0 +1,27 @@
+"""PR pipeline read model (plane ④) for dashboard /pipeline."""
+
+from factory.nats.pipeline.cf_registry import (
+    pages_branch_matches,
+    resolve_pages_project,
+    resolve_pages_repo,
+)
+from factory.nats.pipeline.models import (
+    DEFAULT_REPO,
+    M1_DEPLOY_QUORUM,
+    PipelineCheckRow,
+    PipelineRunRow,
+    PipelineStageStatus,
+)
+from factory.nats.pipeline.store import PipelineStore
+
+__all__ = [
+    "DEFAULT_REPO",
+    "M1_DEPLOY_QUORUM",
+    "PipelineCheckRow",
+    "PipelineRunRow",
+    "PipelineStageStatus",
+    "PipelineStore",
+    "pages_branch_matches",
+    "resolve_pages_project",
+    "resolve_pages_repo",
+]
