+++ src/factory/nats/pipeline/cf_registry.py
