+++ src/factory/nats/pipeline/models.py
